@@ -4,11 +4,10 @@ import io.github.thebusybiscuit.exoticgarden.items.CustomFood;
 import io.github.thebusybiscuit.slimefun4.api.items.SlimefunItem;
 import io.github.thebusybiscuit.slimefun4.api.player.PlayerBackpack;
 import io.github.thebusybiscuit.slimefun4.api.player.PlayerProfile;
-import io.github.thebusybiscuit.slimefun4.utils.SlimefunUtils;
 import me.profelements.dynatech.DynaTech;
-import me.profelements.dynatech.DynaTechItems;
 import me.profelements.dynatech.events.PicnicBasketFeedPlayerEvent;
 import me.profelements.dynatech.items.backpacks.PicnicBasket;
+
 import org.bukkit.Material;
 import org.bukkit.Sound;
 import org.bukkit.entity.Player;
@@ -16,52 +15,53 @@ import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.entity.EntityDamageEvent;
 import org.bukkit.event.entity.FoodLevelChangeEvent;
+import org.bukkit.event.entity.EntityDamageEvent.DamageCause;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
 
+import java.util.HashMap;
+import java.util.Map;
+
 import javax.annotation.Nonnull;
+import javax.annotation.ParametersAreNonnullByDefault;
 
 public class PicnicBasketListener implements Listener {
 
-    private final DynaTech plugin;
+    private final DynaTech plugin;  
     private final PicnicBasket picnicBasket;
-
-    public PicnicBasketListener(@Nonnull DynaTech plugin, @Nonnull PicnicBasket picnicBasket) {
+    
+    @ParametersAreNonnullByDefault
+    public PicnicBasketListener(DynaTech plugin,PicnicBasket picnicBasket) {
         plugin.getServer().getPluginManager().registerEvents(this, plugin);
 
         this.plugin = plugin;
         this.picnicBasket = picnicBasket;
     }
-
-    @EventHandler
+    
+    @EventHandler 
     public void onHungerLoss(FoodLevelChangeEvent e) {
-        if (e.getEntity() instanceof Player) {
-            Player p = (Player) e.getEntity();
-            if (e.getFoodLevel() < p.getFoodLevel()) {
-                checkAndConsume((Player) e.getEntity());
-            }
+        if (e.getEntity() instanceof Player player && e.getFoodLevel() < player.getFoodLevel()) {
+            checkAndConsume(player);
         }
     }
 
-    @EventHandler
+    @EventHandler 
     public void onHungerDamage(EntityDamageEvent e) {
-        if (e.getEntity() instanceof Player && e.getCause() == EntityDamageEvent.DamageCause.STARVATION) {
-            checkAndConsume((Player) e.getEntity());
+        if (e.getEntity() instanceof Player player && e.getCause() == DamageCause.STARVATION) {
+            checkAndConsume(player);
         }
-    }
+    }       
 
-    private void checkAndConsume(Player p) {
+    private void checkAndConsume(@Nonnull Player p) {
         if (picnicBasket == null || picnicBasket.isDisabled()) {
-            return;
+            return; 
         }
 
         for (ItemStack item : p.getInventory().getContents()) {
-            if (item != null && item.getType() == picnicBasket.getItem().getType() && item.hasItemMeta() && (picnicBasket.isItem(item) || DynaTechItems.SOULBOUND_PICNIC_BASKET.getItem().isItem(item))) {
-                if (SlimefunUtils.canPlayerUseItem(p, picnicBasket.getItem(), true)) {
-                    takeFoodFromPicnicBasket(p, item);
-                } else {
-                    return;
-                }
+            if (picnicBasket.isItem(item) && picnicBasket.canUse(p, true)) {
+                takeFoodFromPicnicBasket(p, item);
+            } else {
+                return;
             }
         }
     }
@@ -74,107 +74,93 @@ public class PicnicBasketListener implements Listener {
         });
 
     }
-
-    private void consumeFood(Player p, ItemStack picnicBasketItem, PlayerBackpack backpack) {
+    
+    private boolean consumeFood(@Nonnull Player p, @Nonnull ItemStack picnicBasketItem, @Nonnull PlayerBackpack backpack) {
         Inventory inv = backpack.getInventory();
         int slot = -1;
 
         for (int i = 0; i < inv.getSize(); i++) {
             ItemStack item = inv.getItem(i);
-
-            if (item != null) {
+            
+            if (item != null && (getFoodItems().keySet().contains(item.getType()) || (DynaTech.isExoticGardenInstalled() && SlimefunItem.getByItem(item) instanceof CustomFood))) {
                 slot = i;
-            }
+            } 
         }
 
         if (slot >= 0) {
-            ItemStack item = inv.getItem(slot);
-            PicnicBasketFeedPlayerEvent event = new PicnicBasketFeedPlayerEvent(p, picnicBasket, picnicBasketItem,
-                    item);
+            ItemStack item = inv.getItem(slot); 
+            PicnicBasketFeedPlayerEvent event = new PicnicBasketFeedPlayerEvent(p, picnicBasket, picnicBasketItem, item); 
             plugin.getServer().getPluginManager().callEvent(event);
 
             if (!event.isCancelled()) {
-                boolean itemConsumed = false;
-                
-                SlimefunItem sfItem = SlimefunItem.getByItem(item);
-                if (DynaTech.isExoticGardenInstalled() && sfItem != null) {
-                    if (sfItem instanceof CustomFood) {
-                        CustomFood cfItem = (CustomFood) sfItem;
-                        if (cfItem.getFoodValue() + p.getFoodLevel() <= 20) {
-                            p.setFoodLevel(p.getFoodLevel() + cfItem.getFoodValue());
-                            itemConsumed = true;
-                        }
-                    }
-                } else {
-                    Material material = item.getType();
-                    if (material == Material.COOKED_PORKCHOP || material == Material.PUMPKIN_PIE
-                            || material == Material.COOKED_BEEF) {
-
-                        if (p.getFoodLevel() <= 12) {
-                            p.setFoodLevel(p.getFoodLevel() + 8);
-                            itemConsumed = true;
-                        }
-                    } else if (material == Material.COOKED_CHICKEN || material == Material.COOKED_MUTTON
-                            || material == Material.COOKED_SALMON || material == Material.GOLDEN_CARROT) {
-
-                        if (p.getFoodLevel() <= 14) {
-                            p.setFoodLevel(p.getFoodLevel() + 6);
-                            itemConsumed = true;
-                        }
-
-                    } else if (material == Material.BAKED_POTATO || material == Material.COOKED_RABBIT
-                            || material == Material.COOKED_COD || material == Material.BREAD) {
-                        if (p.getFoodLevel() <= 15) {
-                            p.setFoodLevel(p.getFoodLevel() + 5);
-                            itemConsumed = true;
-                        }
-
-                    } else if (material == Material.APPLE) {
-
-                        if (p.getFoodLevel() <= 16) {
-                            p.setFoodLevel(p.getFoodLevel() + 4);
-                            itemConsumed = true;
-                        }
-                    } else if (material == Material.CARROT || material == Material.BEEF || material == Material.PORKCHOP
-                            || material == Material.RABBIT) {
-
-                        if (p.getFoodLevel() <= 17) {
-                            p.setFoodLevel(p.getFoodLevel() + 3);
-                            itemConsumed = true;
-                        }
-                    } else if (material == Material.COOKIE || material == Material.MELON_SLICE
-                            || material == Material.CHICKEN || material == Material.COD || material == Material.MUTTON
-                            || material == Material.SALMON || material == Material.SWEET_BERRIES
-                            || material == Material.GLOW_BERRIES ) {
-
-                        if (p.getFoodLevel() <= 18) {
-                            p.setFoodLevel(p.getFoodLevel() + 2);
-                            itemConsumed = true;
-                        }
-                    } else if (material == Material.BEETROOT || material == Material.DRIED_KELP
-                            || material == Material.POTATO || material == Material.TROPICAL_FISH) {
-
-                        if (p.getFoodLevel() <= 19) {
-                            p.setFoodLevel(p.getFoodLevel() + 1);
-                            itemConsumed = true;
-                        }
+                if (getFoodItems().keySet().contains(event.getItemConsumed().getType())) {
+                    p.setFoodLevel(p.getFoodLevel() + getFoodItems().get(event.getItemConsumed().getType())); 
+                    p.setSaturation(p.getSaturation() + 4F);
+                    p.playSound(p.getLocation(), Sound.ENTITY_PLAYER_BURP, 1F, 1F);
+                    
+                    
+                    if (item.getAmount() > 1) {
+                        item.setAmount(item.getAmount() - 1); 
+                    } else {
+                        inv.setItem(slot, null); 
                     }
 
-                }
-
-                if (item.getAmount() > 1 && itemConsumed) {
-                    item.setAmount(item.getAmount() - 1);
-                    p.playSound(p.getLocation(), Sound.ENTITY_PLAYER_BURP, 1F, 1F);
+                    backpack.markDirty();
+                    return true;    
+                } else if (DynaTech.isExoticGardenInstalled() && SlimefunItem.getByItem(item) instanceof CustomFood customFood) {
+                    p.setFoodLevel(p.getFoodLevel() + customFood.getFoodValue()); 
                     p.setSaturation(p.getSaturation() + 4F);
+                    p.playSound(p.getLocation(), Sound.ENTITY_PLAYER_BURP, 1F, 1F); 
 
-                } else if (itemConsumed) {
-                    inv.setItem(slot, null);
-                    p.playSound(p.getLocation(), Sound.ENTITY_PLAYER_BURP, 1F, 1F);
-                    p.setSaturation(p.getSaturation() + 4F);
+                    if (item.getAmount() > 1) {
+                        item.setAmount(item.getAmount() - 1);
+                    } else {
+                        inv.setItem(slot, null);
+                    }
+
+                   backpack.markDirty();
+                   return true;
+                } else { 
+                    //TODO: THIS SHOULD NEVER HAPPEN IF IT DOES SOMETHING DID A BIG NO NO
+                    return false; 
                 }
-
-                backpack.markDirty();
             }
         }
+        return false;
     }
+
+    private Map<Material, Integer> getFoodItems() {
+        Map<Material, Integer> foods = new HashMap<Material, Integer>();
+                
+        foods.put(Material.COOKED_PORKCHOP, 8);
+        foods.put(Material.PUMPKIN_PIE, 8);
+        foods.put(Material.COOKED_BEEF, 8);
+        foods.put(Material.COOKED_CHICKEN, 6);
+        foods.put(Material.COOKED_MUTTON, 6);
+        foods.put(Material.COOKED_SALMON, 6);
+        foods.put(Material.GOLDEN_CARROT, 6);
+        foods.put(Material.BAKED_POTATO, 5);
+        foods.put(Material.COOKED_RABBIT, 5);
+        foods.put(Material.COOKED_COD, 5);
+        foods.put(Material.BREAD, 5);
+        foods.put(Material.APPLE, 4);
+        foods.put(Material.CARROT, 3);
+        foods.put(Material.BEEF, 3);
+        foods.put(Material.PORKCHOP, 3);
+        foods.put(Material.RABBIT, 3);
+        foods.put(Material.COOKIE, 2);
+        foods.put(Material.MELON_SLICE, 2);
+        foods.put(Material.CHICKEN, 2);
+        foods.put(Material.COD, 2);
+        foods.put(Material.MUTTON, 2);
+        foods.put(Material.SALMON, 2);
+        foods.put(Material.SWEET_BERRIES, 2);
+        foods.put(Material.GLOW_BERRIES, 2);
+        foods.put(Material.BEETROOT, 1);
+        foods.put(Material.DRIED_KELP, 1);
+        foods.put(Material.POTATO, 1);
+        foods.put(Material.TROPICAL_FISH, 1);
+
+        return foods;
+    } 
 }
